chore: stash work on muzero torch model
