+++ libraries/mathy_python/mathy/agents/muzero/muzero.py
@@ -3,6 +3,7 @@ from dataclasses import dataclass, field
 from typing import Dict, Generic, List, NamedTuple, Optional, TypeVar
 
 import numpy
+import math
 
 from ...envs.poly_simplify import PolySimplify
 from ...state import (
@@ -13,283 +14,18 @@ from ...state import (
     MathyInputsType,
 )
 from ...state import observations_to_window
-from ..mathy_model import MathyModel
+from .network import Network, NetworkOutput
+from .types import (
+    Environment,
+    ActionHistory,
+    Action,
+    Node,
+    Player,
+    MuZeroConfig,
+    MinMaxStats,
+)
 
-Environment = PolySimplify
-
-MAXIMUM_FLOAT_VALUE = float("inf")
-
-KnownBounds = collections.namedtuple("KnownBounds", ["min", "max"])
-
-
-class Player(MathyAgentState):
-    def __init__(self):
-        super(Player, self).__init__(10, "4x+2x", -1)
-
-
-class MinMaxStats(object):
-    """A class that holds the min-max values of the tree."""
-
-    maximum: float
-    minimum: float
-
-    def __init__(self, known_bounds: Optional[KnownBounds]):
-        self.maximum = known_bounds.max if known_bounds else -MAXIMUM_FLOAT_VALUE
-        self.minimum = known_bounds.min if known_bounds else MAXIMUM_FLOAT_VALUE
-
-    def update(self, value: float):
-        self.maximum = max(self.maximum, value)
-        self.minimum = min(self.minimum, value)
-
-    def normalize(self, value: float) -> float:
-        if self.maximum > self.minimum:
-            # We normalize only when we have set the maximum and minimum values.
-            return (value - self.minimum) / (self.maximum - self.minimum)
-        return value
-
-
-class MuZeroConfig(object):
-    def __init__(
-        self,
-        model_width: int,  # TODO: remove this? hack to pad sequences when making observations
-        action_space_size: int,
-        max_moves: int,
-        discount: float,
-        dirichlet_alpha: float,
-        num_simulations: int,
-        batch_size: int,
-        td_steps: int,
-        num_actors: int,
-        lr_init: float,
-        lr_decay_steps: float,
-        visit_softmax_temperature_fn,
-        known_bounds: Optional[KnownBounds] = None,
-    ):
-        ### Self-Play
-        self.action_space_size = action_space_size
-        self.num_actors = num_actors
-
-        self.visit_softmax_temperature_fn = visit_softmax_temperature_fn
-        self.max_moves = max_moves
-        self.num_simulations = num_simulations
-        self.discount = discount
-
-        # Root prior exploration noise.
-        self.root_dirichlet_alpha = dirichlet_alpha
-        self.root_exploration_fraction = 0.25
-
-        # UCB formula
-        self.pb_c_base = 19652
-        self.pb_c_init = 1.25
-
-        # If we already have some information about which values occur in the
-        # environment, we can use them to initialize the rescaling.
-        # This is not strictly necessary, but establishes identical behaviour to
-        # AlphaZero in board games.
-        self.known_bounds = known_bounds
-
-        ### Training
-        self.training_steps = int(1000e3)
-        self.checkpoint_interval = int(1e3)
-        self.window_size = int(1e6)
-        self.batch_size = batch_size
-        self.num_unroll_steps = 5
-        self.td_steps = td_steps
-
-        self.weight_decay = 1e-4
-        self.momentum = 0.9
-
-        # Exponential learning rate schedule
-        self.lr_init = lr_init
-        self.lr_decay_rate = 0.1
-        self.lr_decay_steps = lr_decay_steps
-        self.model_width = model_width
-
-    def new_game(self, env: Environment = None):
-        return Game(self.action_space_size, self.discount, env, self.model_width)
-
-
-@dataclass
-class Action(object):
-    index: int
-
-    def __hash__(self):
-        return self.index
-
-    def __eq__(self, other):
-        return self.index == other.index
-
-    def __gt__(self, other):
-        return self.index > other.index
-
-
-@dataclass
-class Node(object):
-    prior: float
-    visit_count: int = 0
-    to_play: int = 0
-    value_sum: float = 0
-    children: Dict[str, "Node"] = field(default_factory=dict)
-    hidden_state: Optional[List[float]] = None
-    reward: float = 0.0
-
-    def expanded(self) -> bool:
-        return len(self.children) > 0
-
-    def value(self) -> float:
-        if self.visit_count == 0:
-            return 0
-        return self.value_sum / self.visit_count
-
-
-class ActionHistory(object):
-    """MCTS history container used inside the search.
-
-  Only used to keep track of the actions executed.
-  """
-
-    def __init__(self, history: List[Action], action_space_size: int):
-        self.history = list(history)
-        self.action_space_size = action_space_size
-
-    def clone(self):
-        return ActionHistory(self.history, self.action_space_size)
-
-    def add_action(self, action: Action):
-        self.history.append(action)
-
-    def last_action(self) -> Action:
-        return self.history[-1]
-
-    def action_space(self) -> List[Action]:
-        return [Action(i) for i in range(self.action_space_size)]
-
-    def to_play(self) -> MathyAgentState:
-        return MathyAgentState(-1, "4x+2x", -1)
-
-
-class NetworkOutput(NamedTuple):
-    value: float
-    reward: float
-    policy_logits: Dict[Action, float]
-    hidden_state: List[float]
-
-
-class Network(MathyModel):
-    def initial_inference(self, image) -> NetworkOutput:
-        # representation + prediction function
-        out = self(image)
-        return out
-
-    def recurrent_inference(self, hidden_state, action) -> NetworkOutput:
-        # dynamics + prediction function
-        return NetworkOutput(0, 0, {}, [])
-
-    def get_weights(self):
-        # Returns the weights of this network.
-        return []
-
-    def training_steps(self) -> int:
-        # How many steps / batches the network has been trained for.
-        return 0
-
-
-class Game(object):
-    """A single episode of interaction with the environment."""
-
-    def __init__(
-        self,
-        action_space_size: int,
-        discount: float,
-        env: Environment,
-        model_width: int,
-    ):
-        self.environment = env
-        self.state, self.problem = self.environment.get_initial_state()
-        self.history = []
-        self.rewards = []
-        self.child_visits = []
-        self.root_values = []
-        self.action_space_size = action_space_size
-        self.discount = discount
-        self.done = False
-        self.model_width = model_width
-
-    def terminal(self) -> bool:
-        return self.done is True
-
-    def legal_actions(self, current_observation) -> List[Action]:
-        # Game specific calculation of legal actions.
-        mask = self.environment.get_valid_moves(self.state)
-        return [Action(index) for index, valid in enumerate(mask) if valid == 1]
-
-    def apply(self, action: Action):
-        observation, reward, done, info = self.environment.step(action)
-        self.done = done
-        self.rewards.append(reward)
-        self.history.append(action)
-
-    def store_search_statistics(self, root: Node):
-        sum_visits = sum(child.visit_count for child in root.children.values())
-        action_space = (Action(index) for index in range(self.action_space_size))
-        self.child_visits.append(
-            [
-                root.children[a].visit_count / sum_visits if a in root.children else 0
-                for a in action_space
-            ]
-        )
-        self.root_values.append(root.value())
-
-    def step(self, action):
-        self.state, transition, change = self.mathy.get_next_state(self.state, action)
-        observation = self._observe(self.state)
-        info = {"transition": transition}
-        done = is_terminal_transition(transition)
-        self.last_action = action
-        self.last_change = change
-        self.last_reward = round(float(transition.reward), 4)
-        return observation, transition.reward, done, info
-
-    def make_image(self, state_index: int) -> MathyInputsType:
-        obs = self.environment.state_to_observation(self.state, rnn_size=1)
-        return observations_to_window([obs], self.model_width).to_inputs()
-
-    def make_target(
-        self, state_index: int, num_unroll_steps: int, td_steps: int, to_play: Player
-    ):
-        # The value target is the discounted root value of the search tree N steps
-        # into the future, plus the discounted sum of all rewards until then.
-        targets = []
-        for current_index in range(state_index, state_index + num_unroll_steps + 1):
-            bootstrap_index = current_index + td_steps
-            if bootstrap_index < len(self.root_values):
-                value = self.root_values[bootstrap_index] * self.discount ** td_steps
-            else:
-                value = 0
-
-            for i, reward in enumerate(self.rewards[current_index:bootstrap_index]):
-                value += (
-                    reward * self.discount ** i
-                )  # pytype: disable=unsupported-operands
-
-            if current_index < len(self.root_values):
-                targets.append(
-                    (
-                        value,
-                        self.rewards[current_index],
-                        self.child_visits[current_index],
-                    )
-                )
-            else:
-                # States past the end of games are treated as absorbing states.
-                targets.append((0, 0, []))
-        return targets
-
-    def to_play(self) -> Player:
-        return Player()
-
-    def action_history(self) -> ActionHistory:
-        return ActionHistory(self.history, self.action_space_size)
+from .game import Game
 
 
 # Each game is produced by starting at the initial board position, then
