+++ libraries/mathy_python/mathy/agents/muzero/types.py
@@ -1,0 +1,158 @@
+import collections
+from dataclasses import dataclass, field
+from typing import Dict, List, Optional
+from ...env import MathyEnv
+
+Environment = MathyEnv
+
+MAXIMUM_FLOAT_VALUE = float("inf")
+
+KnownBounds = collections.namedtuple("KnownBounds", ["min", "max"])
+
+
+@dataclass
+class Player:
+    id: int = 1
+
+
+class MinMaxStats(object):
+    """A class that holds the min-max values of the tree."""
+
+    maximum: float
+    minimum: float
+
+    def __init__(self, known_bounds: Optional[KnownBounds]):
+        self.maximum = known_bounds.max if known_bounds else -MAXIMUM_FLOAT_VALUE
+        self.minimum = known_bounds.min if known_bounds else MAXIMUM_FLOAT_VALUE
+
+    def update(self, value: float):
+        self.maximum = max(self.maximum, value)
+        self.minimum = min(self.minimum, value)
+
+    def normalize(self, value: float) -> float:
+        if self.maximum > self.minimum:
+            # We normalize only when we have set the maximum and minimum values.
+            return (value - self.minimum) / (self.maximum - self.minimum)
+        return value
+
+
+class MuZeroConfig(object):
+    def __init__(
+        self,
+        model_width: int,  # TODO: remove this? hack to pad sequences when making observations
+        action_space_size: int,
+        max_moves: int,
+        discount: float,
+        dirichlet_alpha: float,
+        num_simulations: int,
+        batch_size: int,
+        td_steps: int,
+        num_actors: int,
+        lr_init: float,
+        lr_decay_steps: float,
+        visit_softmax_temperature_fn,
+        known_bounds: Optional[KnownBounds] = None,
+    ):
+        ### Self-Play
+        self.action_space_size = action_space_size
+        self.num_actors = num_actors
+
+        self.visit_softmax_temperature_fn = visit_softmax_temperature_fn
+        self.max_moves = max_moves
+        self.num_simulations = num_simulations
+        self.discount = discount
+
+        # Root prior exploration noise.
+        self.root_dirichlet_alpha = dirichlet_alpha
+        self.root_exploration_fraction = 0.25
+
+        # UCB formula
+        self.pb_c_base = 19652
+        self.pb_c_init = 1.25
+
+        # If we already have some information about which values occur in the
+        # environment, we can use them to initialize the rescaling.
+        # This is not strictly necessary, but establishes identical behaviour to
+        # AlphaZero in board games.
+        self.known_bounds = known_bounds
+
+        ### Training
+        self.training_steps = int(1000e3)
+        self.checkpoint_interval = int(1e3)
+        self.window_size = int(1e6)
+        self.batch_size = batch_size
+        self.num_unroll_steps = 5
+        self.td_steps = td_steps
+
+        self.weight_decay = 1e-4
+        self.momentum = 0.9
+
+        # Exponential learning rate schedule
+        self.lr_init = lr_init
+        self.lr_decay_rate = 0.1
+        self.lr_decay_steps = lr_decay_steps
+        self.model_width = model_width
+
+    def new_game(self, env: Environment = None):
+        from .game import Game
+
+        return Game(self.action_space_size, self.discount, env, self.model_width)
+
+
+@dataclass
+class Action(object):
+    index: int
+
+    def __hash__(self):
+        return self.index
+
+    def __eq__(self, other):
+        return self.index == other.index
+
+    def __gt__(self, other):
+        return self.index > other.index
+
+
+@dataclass
+class Node(object):
+    prior: float
+    visit_count: int = 0
+    to_play: int = 0
+    value_sum: float = 0
+    children: Dict[str, "Node"] = field(default_factory=dict)
+    hidden_state: Optional[List[float]] = None
+    reward: float = 0.0
+
+    def expanded(self) -> bool:
+        return len(self.children) > 0
+
+    def value(self) -> float:
+        if self.visit_count == 0:
+            return 0
+        return self.value_sum / self.visit_count
+
+
+class ActionHistory(object):
+    """MCTS history container used inside the search.
+
+  Only used to keep track of the actions executed.
+  """
+
+    def __init__(self, history: List[Action], action_space_size: int):
+        self.history = list(history)
+        self.action_space_size = action_space_size
+
+    def clone(self):
+        return ActionHistory(self.history, self.action_space_size)
+
+    def add_action(self, action: Action):
+        self.history.append(action)
+
+    def last_action(self) -> Action:
+        return self.history[-1]
+
+    def action_space(self) -> List[Action]:
+        return [Action(i) for i in range(self.action_space_size)]
+
+    def to_play(self) -> int:
+        return 1
